+++ src/components/App.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 
 import { v4 as uuidv4 } from "uuid";
 
@@ -34,6 +34,7 @@ const App = () => {
 
   let layoutId = useRef(uuidv4())
   let savedLayouts = useRef([])
+  let layout = useRef(cells)
 
   const handleSave = () => {
     const index = savedLayouts.current.findIndex((layout) => layout.id === layoutId.current)
@@ -102,6 +103,93 @@ const App = () => {
     ]);
   }
 
+  const handleDropImage = (item, id) => {
+    const updatedCells = layout.current.map((cell) => cell);
+    const dropIndex = updatedCells.findIndex((cell) => cell.id === id);
+    updatedCells[dropIndex].src = item.imageSrc;
+
+    setCells(updatedCells)
+    layout.current = updatedCells;
+  }
+  
+  const handleSwapContent = (dragId, dropId) => {
+
+    // Create a deep copy of the cells array to avoid modifying the state directly.
+    const updatedCells = layout.current.map((cell) => cell);
+
+    const dragIndex = updatedCells.findIndex((cell) => cell.id === dragId);
+    const dropIndex = updatedCells.findIndex((cell) => cell.id === dropId);
+
+    if (dragIndex !== -1 && dropIndex !== -1) {
+      // Swap the src property between the two cells.
+      const dragSrc = updatedCells[dragIndex].src;
+      updatedCells[dragIndex].src = updatedCells[dropIndex].src;
+      updatedCells[dropIndex].src = dragSrc;
+
+      // Set the updated cells array to the state.
+      setCells(updatedCells);
+      setSelectedCell(null);
+    }
+  }
+
+  const handleVerticalSplit = (selectedCell) => {
+    const updatedCells = cells.flatMap((cell) => {
+      if (cell.id === selectedCell.id) {
+        const splitWidth = cell.w / 2;
+        const newCell1 = { id: uuidv4(), w: splitWidth, h: cell.h, src: selectedCell.src };
+        const newCell2 = { id: uuidv4(), w: splitWidth, h: cell.h, src: selectedCell.src };
+        return [newCell1, newCell2];
+      }
+      return cell;
+    });
+
+    setCells(updatedCells);
+    layout.current = updatedCells;
+  };
+
+  const handleHorizontalSplit = (selectedCell) => {
+
+      let colsAvailable = gridCols;
+      let newCell2 = null;
+      let placedFirstCell = false;
+
+      const updatedCells = cells.flatMap((cell, i, arr) => {
+        if (cell.id === selectedCell.id) {
+          const splitHeight = cell.h / 2;
+          const newCell1 = { id: uuidv4(), w: cell.w, h: splitHeight, src: selectedCell.src };
+          newCell2 = { id: uuidv4(), w: cell.w, h: splitHeight, src: selectedCell.src };
+
+          if (i === arr.length - 1) return [newCell1, newCell2];
+          placedFirstCell = true;
+          colsAvailable -= newCell1.w;
+          if (colsAvailable <= 0) colsAvailable = gridCols;
+          return newCell1;
+        } else if (placedFirstCell && colsAvailable === gridCols) {
+          placedFirstCell = null;
+          return [newCell2, cell]
+        } else {
+          colsAvailable -= cell.w;
+          if (colsAvailable <= 0) colsAvailable = gridCols;
+          return cell;
+        }
+      });
+    
+      setCells(updatedCells);
+      layout.current = updatedCells
+  };
+
+  useEffect(() => {
+    if(!splitCell) return;
+    if(!selectedCell) return;
+    
+    if(splitCell === "vertical") handleVerticalSplit(selectedCell)
+    if(splitCell === "horizontal") handleHorizontalSplit(selectedCell)
+    
+    setSplitCell(null);
+    setSelectedCell(null);
+    // eslint-disable-next-line
+  }, [splitCell]);
+
   return (
     <>
       <ActionArea 
@@ -117,6 +205,8 @@ const App = () => {
         <LayoutArea 
           cells={cells}
           gridCols={gridCols}
+          handleDropImage={handleDropImage}
+          handleSwapContent={handleSwapContent}
           selectedCell = {selectedCell}
           setCells={setCells}
           setEditView={setEditView}
